bugfix: density was init as int in getter function

+++ pedestrian/StartDistribution.h
@@ -44,7 +44,7 @@ class StartDistribution
 private:
      int _roomID;
      int _nPeds;
-     int _DPeds;
+     double _DPeds;
      int _groupID;
      int _goalID;
      int _routerID;
